refactor: Separate user-facing Processor trait from generated internals
Improves developer experience by clearly distinguishing between:
- `Processor` trait: User-implemented behavior (setup, teardown, process)
- `GeneratedProcessor` trait: Macro-generated internals (doc-hidden)

Key changes:
- New `__generated_private` module for internal traits
- `Processor` trait now minimal (3 methods, only process() required)
- All processors updated to `impl crate::core::Processor for X::Processor`
- Helper methods moved to separate impl blocks
- Added module-level convenience: `MyProcessor::node()` and `MyProcessor::Config`

The new pattern:
```rust
impl streamlib::Processor for MyProcessor::Processor {
    fn process(&mut self) -> Result<()> { ... }
}

// Convenience API
MyProcessor::node(MyProcessor::Config { ... })
```
