+++ libs/streamlib/src/core/graph/graph_tests.rs
@@ -42,7 +42,7 @@ struct MockProcessor {
     config: MockConfig,
 }
 
-impl MockProcessor::Processor {
+impl crate::core::Processor for MockProcessor::Processor {
     fn setup(
         &mut self,
         _ctx: &crate::core::context::RuntimeContext,
@@ -50,6 +50,9 @@ impl MockProcessor::Processor {
         Ok(())
     }
     fn teardown(&mut self) -> crate::core::error::Result<()> {
+        Ok(())
+    }
+    fn process(&mut self) -> crate::core::error::Result<()> {
         Ok(())
     }
 }
@@ -67,7 +70,7 @@ struct MockOutputOnlyProcessor {
     config: MockConfig,
 }
 
-impl MockOutputOnlyProcessor::Processor {
+impl crate::core::Processor for MockOutputOnlyProcessor::Processor {
     fn setup(
         &mut self,
         _ctx: &crate::core::context::RuntimeContext,
@@ -75,6 +78,9 @@ impl MockOutputOnlyProcessor::Processor {
         Ok(())
     }
     fn teardown(&mut self) -> crate::core::error::Result<()> {
+        Ok(())
+    }
+    fn process(&mut self) -> crate::core::error::Result<()> {
         Ok(())
     }
 }
@@ -92,7 +98,7 @@ struct MockInputOnlyProcessor {
     config: MockConfig,
 }
 
-impl MockInputOnlyProcessor::Processor {
+impl crate::core::Processor for MockInputOnlyProcessor::Processor {
     fn setup(
         &mut self,
         _ctx: &crate::core::context::RuntimeContext,
@@ -100,6 +106,9 @@ impl MockInputOnlyProcessor::Processor {
         Ok(())
     }
     fn teardown(&mut self) -> crate::core::error::Result<()> {
+        Ok(())
+    }
+    fn process(&mut self) -> crate::core::error::Result<()> {
         Ok(())
     }
 }
