+++ libs/streamlib/src/core/processors/audio_channel_converter.rs
@@ -46,7 +46,7 @@ pub struct AudioChannelConverterProcessor {
     frame_counter: u64,
 }
 
-impl AudioChannelConverterProcessor::Processor {
+impl crate::core::Processor for AudioChannelConverterProcessor::Processor {
     fn setup(&mut self, _ctx: &RuntimeContext) -> Result<()> {
         tracing::info!(
             "[AudioChannelConverter] setup() - mode: {:?}",
